# Conflicts:
#	app/src/main/res/layout/act_sign_up.xml

+++ app/src/main/java/com/ufrpe/feelingsbox/dominio/usuario/Usuario.java
@@ -1,14 +1,10 @@
 package com.ufrpe.feelingsbox.dominio.usuario;
 
-/**
- * Created by evele on 18/07/2017.
- */
-
 public class Usuario {
     private String email;
     private String senha;
     private String nick;
-    private int id;
+    private long id;
 
     public String getEmail(){
 
@@ -38,12 +34,12 @@ public class Usuario {
         this.nick = novoNick;
     }
 
-    public int getId(){
+    public long getId(){
 
         return this.id;
     }
 
-    public void setId(int novoId){
+    public void setId(long novoId){
 
         this.id = novoId;
     }
